+++ spec/features/admin_views_completed_claim_spec.rb
@@ -4,13 +4,14 @@ RSpec.feature "Admin can view completed claims" do
   before { @signed_in_user = sign_in_as_service_operator }
 
   scenario "Viewing a claim that has a decision made " do
-    claim_with_decision = create(:claim, :approved)
+    claim_with_decision = create(:claim, :approved, policy: MathsAndPhysics)
 
     visit admin_claim_tasks_path(claim_with_decision)
 
     within("span#claim-heading") do
       expect(page).to have_content("Approved")
     end
+    expect(page).to have_content("Claim amount £2,000.00")
   end
 
   scenario "Viewing a claim that does not have decision made" do
@@ -23,7 +24,7 @@ RSpec.feature "Admin can view completed claims" do
     end
   end
 
-  scenario "Viewing a payroll status for an approved claim which has a payment" do
+  scenario "Viewing an approved claim which has a payment" do
     payroll_run = create(:payroll_run, claims_counts: {StudentLoans => 1})
     payroll_run_date = payroll_run.created_at.strftime("%B %Y")
     claim_with_payment = payroll_run.claims.first