+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog]
 ## [Unreleased]
 
 - Display payroll status of claim within task view
+- Display claim amount within task view
 
 ## [Release 073] - 2020-04-07